Move into GalaxyState
