+++ tests/test_state.py
@@ -1,4 +1,5 @@
 import pytest
+from typing import Dict
 
 from edmfs.event_processors import UnknownStarSystemError
 from edmfs.state import StarSystem, PilotState, Station, GalaxyState, Mission, ResolvingDict
@@ -60,6 +61,24 @@ def test_galaxy_state_init_args():
 def test_galaxy_state_get_system(resolver, star_systems, system_address, expected_star_system):
     galaxy_state = GalaxyState(resolver, star_systems)
     assert galaxy_state.systems.get(system_address, None) == expected_star_system
+
+@pytest.mark.parametrize(
+    "commodity_name, expected_result",
+    [
+        ("Platinum", {"BuyPrice":30908, "SellPrice":30906, "MeanPrice":58263, "StockBracket":0, "DemandBracket":0, "Stock":0}),
+        ("Animal Meat", {"BuyPrice":0, "SellPrice":1864, "MeanPrice":1539, "StockBracket":0, "DemandBracket":3, "Stock":0}),
+        ("Classified Experimental Equipment", {"BuyPrice":4002, "SellPrice":4001, "MeanPrice":11423, "StockBracket":0, "DemandBracket":0, "Stock":0}),
+        ("Not A Commodity", None)
+    ]
+)
+def test_galaxy_state_get_last_market_entry(commodity_name:str, expected_result:Dict):
+    market_entry = GalaxyState().get_last_market_entry(commodity_name, "tests/market_files/market.json")
+    if expected_result:
+        assert market_entry["Name_Localised"] == commodity_name
+        for key in expected_result:
+            assert market_entry[key] == expected_result[key]
+    else:
+        assert not market_entry
 
 def raise_error(x):
     raise UnknownStarSystemError(x)
